backend/AI_backend/requirements.txt
Flask
Flask-CORS
Flask-SQLAlchemy
Flask-Marshmallow
Pillow
tensorflow
numpy
sqlalchemy
marshmallow
python-dateutil
mysql-connector-python
